Realizo cambios en la estructura del proyecto para dividir la logica de la interfaz, tambien cambio el nombre de algunas funciones y reacomodo partes del codigo para intentar limitarlas a hacer su funcion y no mas que eso
