+++ tools/create_hall.py
@@ -1,0 +1,9 @@
+
+def create_halls(rows, columns):
+  """Crea la sala de cine"""
+  
+  seats = []
+
+  for _ in range(rows): seats.append([0] * columns)
+
+  return seats
