+++ tools/disable_seats.py
@@ -1,0 +1,7 @@
+def disable_seats(hall, row, column):
+    """Deshabilita butacas marcándolas con 'X'. Modifica la matriz en el lugar."""
+
+    row_label, col_label = row + 1, column + 1
+
+    hall[row][column] = ' '
+    print(f"Butaca F{row_label}-C{col_label} inhabilitada correctamente.")
